requirements.txt
# ──────────────────────────────────────────────
# Core API
# ──────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# ──────────────────────────────────────────────
# AI / ML — Vector & Semantic Search
# ──────────────────────────────────────────────
sentence-transformers>=2.6.0
faiss-cpu>=1.8.0

# ──────────────────────────────────────────────
# AI / ML — CLIP Validation
# ──────────────────────────────────────────────
open-clip-torch>=2.24.0
torch>=2.0.0
Pillow>=10.0.0

# ──────────────────────────────────────────────
# RAG (optional — comment out if not using LLM)
# ──────────────────────────────────────────────
langchain>=0.1.0

# ──────────────────────────────────────────────
# Wikipedia RAG (optional)
# ──────────────────────────────────────────────
Wikipedia-API>=0.6.0

# ──────────────────────────────────────────────
# Google Cloud (required on Cloud Run)
# ──────────────────────────────────────────────
google-cloud-storage>=2.14.0

# ──────────────────────────────────────────────
# Real Dataset Connectors
# ──────────────────────────────────────────────
requests>=2.31.0            # Sketchfab + Poly Haven HTTP calls
httpx>=0.27.0               # async-ready HTTP (future use)
objaverse>=0.1.7            # Objaverse 800k+ 3D objects (Allen AI)

# ──────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────
numpy>=1.26.0
python-dotenv>=1.0.0